+++ nix/overlay.nix
@@ -1,10 +1,22 @@
 self: super:
 let
   inherit (self.arion-project) haskellPkgs;
+
+  srcDir = ../src; # TODO gitignoreSource + whitelist nix and arion-image
+  eval = import (srcDir + "/nix/eval-composition.nix");
+  build = args@{...}:
+    let composition = eval args;
+    in composition.config.build.dockerComposeYaml;
+  hlib = super.haskell.lib;
 in
 {
+
   arion-v0 = super.callPackage ../arion.nix {};
-  arion = super.haskell.lib.justStaticExecutables haskellPkgs.arion-compose;
+  arion = hlib.justStaticExecutables (hlib.overrideCabal haskellPkgs.arion-compose (o: {
+      passthru = o.passthru // {
+        inherit eval build;
+      };
+    }));
   tests = super.callPackage ../tests {};
   doc = super.callPackage ../doc {};