+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.
+* *BREAKING:* useHostStore now uses a proper empty base image, like `scratch`.